Add placeholder icons for missing quote thumbnails.
Previously, if we were missing a quote thumbnail for images or video,
we'd show nothing. To be consistent with desktop and also have a
fallback strategy for missing icons, I've put in a placeholder for that
scenario.

Doing this involved a refactor of how we set the QuoteView's thumbnail.
It should be cleaner and less bug-prone now.